+++ src/components/Tooltip/tooltipChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Update",
+    affects: ["Styles"],
+    notes: [
+      "Updates text color to `ui.typography.inverse.heading` on light mode.",
+    ],
+  },
+  {
     date: "2024-09-05",
     version: "3.3.1",
     type: "Update",