+++ CHANGELOG.md
@@ -10,6 +10,10 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ### Updates
 
+- Updates `Notification` variant font colors to sync with VDL.
+- Updates `Tooltip` text color.
+- Updates the `SearchBar` component to remove the text input right border.
+- Updates the `SkeletonLoader` component to make the button placeholder look more like a button.
 - Updates the `SubNav` component to add dark mode color styles.
 
 ## 3.6.0 (April 10, 2025)